requirements.txt
numpy
ffmpeg-python==0.2.0
torch
--extra-index-url https://download.pytorch.org/whl/cu113
git+https://github.com/openai/whisper.git